scripts/generate-template.sh: Don't hide errors
As stdout is used for the file, stderr is used by default for progress
indicators. But if we redirect it to /dev/null we will miss any other errors out
there. We can just run with "-q" (quiet) that hides the progress indicators and
have the errors shown as expected.

So this patch does that: run with -q and don't redirect stderr.